requirements.txt
# requirements.txt
streamlit>=1.32.0
langchain-community>=0.0.11
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
playwright>=1.32.0
python-dotenv>=1.0.0